fix(deploy): unset shell DATABASE_URL, block foreign DB, fix windowMs BigInt

+++ site/prisma/migrations/0001_init/migration.sql
@@ -212,7 +212,7 @@ CREATE TABLE "UserMissionDay" (
 CREATE TABLE "LeaderboardRewardRule" (
     "id" TEXT NOT NULL,
     "name" TEXT NOT NULL,
-    "windowMs" INTEGER NOT NULL,
+    "windowMs" BIGINT NOT NULL,
     "rankMin" INTEGER NOT NULL,
     "rankMax" INTEGER NOT NULL,
     "bonusCents" INTEGER NOT NULL DEFAULT 0,
